requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
aiohttp==3.10.5
beautifulsoup4==4.12.3
python-telegram-bot==21.3
schedule==1.2.2
PyYAML==6.0.2
chromadb==0.5.5
tenacity==8.5.0
httpx==0.27.2
pydantic==2.9.2
